requirements.txt
marshmallow==3.21.0
SQLAlchemy==2.0.27
alembic==1.13.1
Flask==2.3.3
Flask-RESTful==0.3.10
flask_jwt_extended==4.6.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
flask_migrate==4.0.7
pymysql==1.1.1
flask_mail==0.10.0